+++ src/sysindep/unistd/chown.c
@@ -2,5 +2,5 @@
 #include <unistd.h>
 
 int chown(const char *path, uid_t owner, gid_t group) {
-    return (int)syscall(SYS_chown, (long)path, owner, group);
+    return (int)syscall(SYS_chown, (long)path, (long)owner, (long)group);
 }
