+++ src/sysindep/unistd/execle.c
@@ -1,4 +1,5 @@
 #include <stdarg.h>
+#include <stdint.h>
 #include <stdlib.h>
 #include <unistd.h>
 
@@ -16,11 +17,11 @@ int execle(const char *path, const char *arg, ...) {
     envp = va_arg(ap, char **);
     va_end(ap);
 
-    argv = (char **)malloc((argc + 1) * sizeof(char *));
+    argv = (char **)malloc(((size_t)argc + 1) * sizeof(char *));
     if(!argv)
         return -1;
 
-    argv[0] = (char *)arg;
+    argv[0] = (char *)(uintptr_t)arg;
     va_start(ap, arg);
     for(i = 1; i <= argc; i++) {
         argv[i] = va_arg(ap, char *);
