+++ src/sysindep/unistd/execlp.c
@@ -1,4 +1,5 @@
 #include <stdarg.h>
+#include <stdint.h>
 #include <stdlib.h>
 #include <unistd.h>
 
@@ -14,11 +15,11 @@ int execlp(const char *file, const char *arg, ...) {
     }
     va_end(ap);
 
-    argv = (char **)malloc((argc + 1) * sizeof(char *));
+    argv = (char **)malloc(((size_t)argc + 1) * sizeof(char *));
     if(!argv)
         return -1;
 
-    argv[0] = (char *)arg;
+    argv[0] = (char *)(uintptr_t)arg;
     va_start(ap, arg);
     for(i = 1; i <= argc; i++) {
         argv[i] = va_arg(ap, char *);
