+++ src/sysindep/unistd/fchown.c
@@ -2,5 +2,5 @@
 #include <unistd.h>
 
 int fchown(int fd, uid_t owner, gid_t group) {
-    return (int)syscall(SYS_fchown, fd, owner, group);
+    return (int)syscall(SYS_fchown, (long)fd, (long)owner, (long)group);
 }
