+++ src/sysindep/unistd/read.c
@@ -3,5 +3,5 @@
 #include <unistd.h>
 
 ssize_t read(int fd, void *buf, size_t count) {
-    return (ssize_t)syscall(SYS_read, fd, (long)buf, count);
+    return (ssize_t)syscall(SYS_read, fd, (long)buf, (long)count);
 }
