+++ src/sysindep/unistd/seteuid.c
@@ -1,5 +1,5 @@
 #include <sys/syscall.h>
 #include <unistd.h>
 int seteuid(uid_t euid) {
-    return (int)syscall(SYS_seteuid, euid);
+    return (int)syscall(SYS_seteuid, (long)euid);
 }
