+++ src/sysindep/unistd/setgid.c
@@ -1,5 +1,5 @@
 #include <sys/syscall.h>
 #include <unistd.h>
 int setgid(gid_t id) {
-    return (int)syscall(SYS_setgid, id);
+    return (int)syscall(SYS_setgid, (long)id);
 }
