+++ src/sysindep/unistd/setuid.c
@@ -1,5 +1,5 @@
 #include <sys/syscall.h>
 #include <unistd.h>
 int setuid(uid_t id) {
-    return (int)syscall(SYS_setuid, id);
+    return (int)syscall(SYS_setuid, (long)id);
 }
